requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Machine Learning and Data Science
scikit-learn>=1.3.0
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.30.0
nltk>=3.8.0
textblob>=0.17.0
vaderSentiment>=3.3.2
yfinance>=0.2.0
newsapi-python>=0.2.6

# Data Processing and Analysis
beautifulsoup4>=4.12.0
feedparser>=6.0.0
tweepy>=4.14.0
praw>=7.7.0
selenium>=4.10.0

# Database and Caching
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.6.0
alembic>=1.11.0

# API and Web Framework
flask>=2.3.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0
flask-migrate>=4.0.0
flask-jwt-extended>=4.5.0
marshmallow>=3.20.0
webargs>=8.2.0

# Real-time Communication
websockets>=11.0.0
socketio>=5.8.0
flask-socketio>=5.3.0

# Monitoring and Logging
prometheus-client>=0.17.0
structlog>=23.1.0
sentry-sdk>=1.28.0

# Task Queue and Scheduling
celery>=5.3.0
redis>=4.6.0
apscheduler>=3.10.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0

# Development Tools
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Telegram Bot
python-telegram-bot>=20.4.0

# Financial and Trading
ccxt>=4.0.0
ta-lib>=0.4.0
quantlib>=1.31.0

# Utilities
click>=8.1.0
python-dateutil>=2.8.0
pytz>=2023.3
tqdm>=4.65.0
joblib>=1.3.0

